fixup: Bug with pushing data and constants moved to correct place
