+++ src/utilities/Constants.ts
@@ -1,1 +1,6 @@
-export const SPACE_CHANNEL_PREFIX = '_ably_space_';
+const SPACE_CHANNEL_PREFIX = '_ably_space_';
+
+const CURSOR_POSITION_EVENT = 'cursorPosition';
+const CURSOR_DATA_EVENT = 'cursorDataUpdate';
+
+export { SPACE_CHANNEL_PREFIX, CURSOR_POSITION_EVENT, CURSOR_DATA_EVENT };
